requirements.txt
django==6.0.2
djangorestframework==3.16.1
pandas==3.0.1
openpyxl==3.1.5
pdfplumber==0.11.9
python-docx==1.2.0
scikit-learn==1.8.0
plotly==6.6.0
whitenoise==6.11.0
python-decouple==3.8
Pillow==12.1.1
numpy==2.4.2
scipy==1.17.0